Factor out new helper class for managing mocking SQL state
